Issue 36060: Can't place order for ordered qty less than 1 with guest checkout

+++ app/code/Magento/Checkout/Model/GuestPaymentInformationManagement.php
@@ -176,7 +176,7 @@ class GuestPaymentInformationManagement implements \Magento\Checkout\Api\GuestPa
         }
         $this->limitShippingCarrier($quote);
 
-        if (!(int)$quote->getItemsQty()) {
+        if (!(float)$quote->getItemsQty()) {
             throw new CouldNotSaveException(__('Some of the products are disabled.'));
         }
